Fixes to change weights to run on backgrounds and send script

+++ bmaker/genfiles/inc/cross_sections.hh
@@ -8,7 +8,7 @@
 namespace xsec{
 
   float crossSection(const TString &file);
-  void signalCrossSection(int glu_mass, double &xsec, double &xsec_unc);
+  void signalCrossSection(int glu_mass, float &xsec, float &xsec_unc);
   float fractionNegWeights(const TString &file);
 }
